pyproject.toml
[tool.poetry]
name = "test-task-itbell"
version = "0.1.0"
description = ""
authors = ["Sergei Demchenko <hb13718@gmail.com>"]
readme = "README.md"
packages = [{include = "test_task_itbell"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
djoser = "^2.2.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
